try systray integration in dwm

+++ compile/dwm-wm/build.sh
@@ -7,21 +7,21 @@ if [ -d "$DIR/dwm" ]; then
     rm -rf $DIR/dwm
 fi
 
-$(svn co https://svn.0mark.unserver.de/dwm/trunk/dwm-sprinkles/ ${DIR}/dwm)
-#$(hg clone http://hg.suckless.org/dwm ${DIR}/dwm)
+$(hg clone http://hg.suckless.org/dwm ${DIR}/dwm)
 cd $DIR
 
 # apply configs
-#cp -f config.h dwm/.
-cp -f sconfig.h dwm/config.h
+cp -f config.h dwm/.
 cp -f config.mk dwm/.
 
 # apply patches
-#patch dwm/dwm.c patches/6.1.trayer.patch.diff
+patch dwm/dwm.c patches/6.1.trayer.patch.diff
 
 # install
 cd dwm
 make clean install
 # copy binary
-cp $DIR/dwm/bin/dwm ~/.dotfiles/bin/dwm
+if [ -f "$DIR/dwm/bin/dwm" ]; then
+    cp $DIR/dwm/bin/dwm ~/.dotfiles/bin/dwm
+fi
